refactor(capabilities): split audio runtime handlers and tests
